refactor(evals): extract shared helpers; drop dead variants module
Eliminate three duplications and one unused module.

- New `tools/evals/lib/load-anthropic-env.ts` — the dotenv + ~/.atlas/.env
  layering + ANTHROPIC_API_KEY guard that all three workspace-chat evals
  copy-pasted. Single callsite per eval: `loadAnthropicEnv("workspace-chat evals")`.
- New `tools/evals/lib/resolve-model.ts` — the `WORKSPACE_CHAT_EVAL_MODEL`
  parser that was copied (with three slightly different error messages) into
  each workspace-chat eval. Now takes the env-var name + fallback so any
  future eval can reuse it.
- New `tools/evals/promptfoo/shared/assertions/not-a-refusal.ts` — the
  12-entry NotARefusal assertion bank previously inlined in
  prompt-interpolation/render.ts. Future suites needing "model didn't punt"
  coverage import `notARefusalAsserts(metric?)`.
- Delete `tools/evals/lib/variants.ts` + `variants.test.ts`. Zero non-test
  callers; the multi-model mechanism that landed (LiteLLM proxy + tier-
  labelled providers.yaml) replaces it. Lib test count goes 83 → 75.

`tests.generated.yaml` is byte-identical before/after.
