+++ tools/evals/agents/workspace-chat/elicitation-behavior.eval.ts
@@ -13,33 +13,18 @@
 //   - On `bypass`: the LLM should proceed with the underlying tool call
 //     without surfacing the bypass to the user (operator-only signal).
 
-import { existsSync } from "node:fs";
 import { readFile } from "node:fs/promises";
-import { join, resolve } from "node:path";
-import process from "node:process";
-import {
-  buildRegistryModelId,
-  isRegistryProvider,
-  type RegistryModelId,
-  registry,
-  traceModel,
-} from "@atlas/llm";
-import { getFridayHome } from "@atlas/utils/paths.server";
+import { resolve } from "node:path";
+import { registry, traceModel } from "@atlas/llm";
 import { stepCountIs, streamText, tool } from "ai";
-import dotenv from "dotenv";
 import { z } from "zod";
 import { AgentContextAdapter } from "../../lib/context.ts";
+import { loadAnthropicEnv } from "../../lib/load-anthropic-env.ts";
 import { type BaseEvalCase, defineEval, type EvalRegistration } from "../../lib/registration.ts";
+import { resolveModelId } from "../../lib/resolve-model.ts";
 import { createScore } from "../../lib/scoring.ts";
 
-dotenv.config();
-const globalAtlasEnv = join(getFridayHome(), ".env");
-if (existsSync(globalAtlasEnv)) {
-  dotenv.config({ path: globalAtlasEnv, override: true });
-}
-if (!process.env.ANTHROPIC_API_KEY) {
-  throw new Error("ANTHROPIC_API_KEY is required to run workspace-chat evals");
-}
+loadAnthropicEnv("workspace-chat evals");
 
 const adapter = new AgentContextAdapter();
 
@@ -148,22 +133,7 @@ const cases: ElicitationCase[] = [
   },
 ];
 
-function resolveModelId(): RegistryModelId {
-  const raw = process.env.WORKSPACE_CHAT_EVAL_MODEL;
-  if (!raw) return "anthropic:claude-sonnet-4-20250514";
-  const colonIdx = raw.indexOf(":");
-  if (colonIdx === -1) {
-    throw new Error(`WORKSPACE_CHAT_EVAL_MODEL must be "provider:model", got "${raw}".`);
-  }
-  const provider = raw.slice(0, colonIdx);
-  const model = raw.slice(colonIdx + 1);
-  if (!isRegistryProvider(provider)) {
-    throw new Error(`Unknown provider "${provider}" in WORKSPACE_CHAT_EVAL_MODEL.`);
-  }
-  return buildRegistryModelId(provider, model);
-}
-
-const MODEL_ID = resolveModelId();
+const MODEL_ID = resolveModelId("WORKSPACE_CHAT_EVAL_MODEL", "anthropic:claude-sonnet-4-20250514");
 
 interface RunOutcome {
   captures: CapturedToolCalls;
